Adding exe file for V1.0 and instable file
