+++ UI-DEV/UI1.0.py
@@ -1,4 +1,3 @@
-
 import wx
 import os
 import sys
