+++ VIT Notes Sorter.iss
@@ -1,0 +1,54 @@
+; Script generated by the Inno Setup Script Wizard.
+; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
+
+#define MyAppName "VIT Notes Sorter"
+#define MyAppVersion "1.0"
+#define MyAppPublisher "60Degrees"
+#define MyAppURL "www.60Degrees.com"
+#define MyAppExeName "VIT_Notes_Sorter.exe"
+
+[Setup]
+; NOTE: The value of AppId uniquely identifies this application. Do not use the same AppId value in installers for other applications.
+; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
+AppId={{CC71CE76-1404-42D8-959F-56F940C50AA9}
+AppName={#MyAppName}
+AppVersion={#MyAppVersion}
+;AppVerName={#MyAppName} {#MyAppVersion}
+AppPublisher={#MyAppPublisher}
+AppPublisherURL={#MyAppURL}
+AppSupportURL={#MyAppURL}
+AppUpdatesURL={#MyAppURL}
+DefaultDirName={autopf}\{#MyAppName}
+DisableProgramGroupPage=yes
+LicenseFile=D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\LICENSE
+InfoBeforeFile=D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\Installation.txt
+InfoAfterFile=D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\Finish.txt
+; Remove the following line to run in administrative install mode (install for all users.)
+PrivilegesRequired=lowest
+OutputDir=D:\Offline_Projects\VIT-Notes-Sorting-master
+OutputBaseFilename=VIT Notes Sorter
+SetupIconFile=D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\icon.ico
+Compression=lzma
+SolidCompression=yes
+WizardStyle=modern
+
+[Languages]
+Name: "english"; MessagesFile: "compiler:Default.isl"
+
+[Tasks]
+Name: "desktopicon"; Description: "{cm:CreateDesktopIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked
+
+[Files]
+Source: "D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\VIT_Notes_Sorter.exe"; DestDir: "{app}"; Flags: ignoreversion
+Source: "D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\Course_List.csv"; DestDir: "{app}"; Flags: ignoreversion
+Source: "D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\icon.png"; DestDir: "{app}"; Flags: ignoreversion
+Source: "D:\Offline_Projects\GIT Repo's\VIT-Notes-Sorting\VIT_Notes_Sorter.exe"; DestDir: "{app}"; Flags: ignoreversion
+; NOTE: Don't use "Flags: ignoreversion" on any shared system files
+
+[Icons]
+Name: "{autoprograms}\{#MyAppName}"; Filename: "{app}\{#MyAppExeName}"
+Name: "{autodesktop}\{#MyAppName}"; Filename: "{app}\{#MyAppExeName}"; Tasks: desktopicon
+
+[Run]
+Filename: "{app}\{#MyAppExeName}"; Description: "{cm:LaunchProgram,{#StringChange(MyAppName, '&', '&&')}}"; Flags: nowait postinstall skipifsilent
+
